Convert Integer to Long for createdIndex, modifiedIndex, etcdIndex and ttl.
Fixes #10

+++ src/main/java/mousio/etcd4j/requests/EtcdKeyPutRequest.java
@@ -71,7 +71,7 @@ public class EtcdKeyPutRequest extends EtcdKeyRequest {
    * @param prevIndex to set on key
    * @return Itself for chaining
    */
-  public EtcdKeyPutRequest prevIndex(int prevIndex) {
+  public EtcdKeyPutRequest prevIndex(long prevIndex) {
     this.requestParams.put("prevIndex", prevIndex + "");
     return this;
   }
